requirements.txt
django
Djangorestframework